chore: removes proxy from demo hooks
