+++ packages/plugin-stripe/demo/package.json
@@ -15,7 +15,8 @@
   "dependencies": {
     "dotenv": "^8.2.0",
     "express": "^4.17.1",
-    "payload": "^1.0.27"
+    "payload": "^1.0.27",
+    "stripe": "^10.10.0"
   },
   "devDependencies": {
     "@types/express": "^4.17.9",
