+++ packages/plugin-stripe/demo/src/mocks/serverModule.js
@@ -1,0 +1,7 @@
+export default {};
+export const updateExistingCustomer = () => { }
+export const syncNewCustomer = () => { }
+export const stripeREST = () => { }
+export const stripeWebhooks = () => { }
+export const stripeProxy = () => { }
+export const Stripe = () => { }
