+++ packages/plugin-stripe/src/index.ts
@@ -3,8 +3,6 @@ import { stripeREST } from './routes/rest';
 import { stripeWebhooks } from './routes/webhooks';
 import { StripeConfig } from './types';
 
-export { stripeProxy } from './utilities/stripeProxy';
-
 const stripe = (stripeConfig: StripeConfig) => (config: Config): Config => {
   return ({
     ...config,
